Added documentation to datasets, all networks and wheel holder
Created documentation for spaghetti dataset. with images and errors possible in the dataset
Added an information file per All networks notebook to discuss results. These are found on Wandb as reports in the projects folders

Added pictures of the new wheel holder and short documentation of the new design an print errors